backend/server/requirements.txt
fastapi==0.95.2
sqlalchemy==2.0.16
pymysql==1.0.3
uvicorn==0.22.0
python-dotenv==0.21.0